add edit post text functionality #23

+++ staff/manuel-barzi/product/api/logic/deletePost.js
@@ -1,7 +1,7 @@
 import { User, Post } from '../data/models.js'
 import { validate, errors } from 'com'
 
-const { SystemError, NotFoundError } = errors
+const { SystemError, NotFoundError, OwnershipError } = errors
 
 const deletePost = (userId, postId) => {
     validate.id(userId, 'userId')
@@ -17,6 +17,8 @@ const deletePost = (userId, postId) => {
         })
         .then(post => {
             if (!post) throw new NotFoundError('post not found')
+
+            if (post.author.toString() !== userId) throw new OwnershipError('user is not author of post')
 
             return Post.deleteOne({ _id: post._id })
                 .catch(error => { throw new SystemError(error.message) })
